requirements.txt
# RCJY Media Generator - Production Dependencies
# Royal Commission for Jubail and Yanbu

# Google AI - SOTA Models (Imagen 4, Nano Banana, Veo 3.1, Gemini TTS)
google-genai>=1.0.0

# Web Framework
streamlit>=1.40.0

# Document Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf>=4.0.0
python-docx>=1.0.0
Pillow>=10.0.0
python-pptx>=1.0.0
openpyxl>=3.1.0

# Utilities
urllib3>=2.0.0
